+++ honeybadger_test.go
@@ -10,6 +10,7 @@ import (
 	"testing"
 
 	"github.com/pborman/uuid"
+	"github.com/stretchr/testify/mock"
 )
 
 var (
@@ -18,6 +19,14 @@ var (
 	requests      []*HTTPRequest
 	defaultConfig = *Config
 )
+
+type MockedHandler struct {
+	mock.Mock
+}
+
+func (h *MockedHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	h.Called()
+}
 
 type HTTPRequest struct {
 	Request *http.Request
@@ -252,6 +261,31 @@ func testNoticePayload(t *testing.T, payload hash) bool {
 		}
 	}
 	return true
+}
+
+func TestHandlerCallsHandler(t *testing.T) {
+	mockHandler := &MockedHandler{}
+	mockHandler.On("ServeHTTP").Return()
+
+	handler := Handler(mockHandler)
+	req, _ := http.NewRequest("GET", "", nil)
+	w := httptest.NewRecorder()
+	handler.ServeHTTP(w, req)
+
+	mockHandler.AssertCalled(t, "ServeHTTP")
+}
+
+func TestMetricsHandlerCallsHandler(t *testing.T) {
+	mockHandler := &MockedHandler{}
+	mockHandler.On("ServeHTTP").Return()
+
+	metricsHandler := MetricsHandler(mockHandler)
+
+	req, _ := http.NewRequest("GET", "", nil)
+	w := httptest.NewRecorder()
+	metricsHandler.ServeHTTP(w, req)
+
+	mockHandler.AssertCalled(t, "ServeHTTP")
 }
 
 func assertMethod(t *testing.T, r *http.Request, method string) {